requirements.txt
numpy==1.17.0
pandas==1.0.0
matplotlib==3.2.1
PyYAML==5.1.2
torch==1.6.0
torchvision==0.7.0
alibi==0.5.4
pytorch-gradcam==0.2.1
shap==0.36.0
